chore: project cleanup, refactor types, and add Windows setup script
- Remove dead Electron build workflow and User Prisma model
- Remove unused babel-plugin-react-compiler and @/* path alias
- Add dist/ to backend .gitignore
- Extract job types to app/types/job.ts and Zod schema to app/lib/validations/job.ts
- Fix React lint warnings: replace setState-in-effect with lazy useState and useRef
- Remove Download App and Chrome Extension buttons from homepage
- Update README and hero text to remove Electron/desktop references
- Add setup.bat for one-click Windows install and launch
