+++ backend/.gitignore
@@ -1,5 +1,4 @@
 node_modules
-# Keep environment variables out of version control
 .env
-
+dist
 /src/generated/prisma
