+++ frontend/app/(app)/jobs/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect, useRef } from 'react';
 import KanbanBoard from '../../components/jobs/KanbanBoard';
 import { useJobs, useCreateJob } from '../../hooks/useJobs';
 import demoJobs from '../../lib/demoJobs';
@@ -10,13 +10,13 @@ const isDemo = process.env.NEXT_PUBLIC_APP_MODE === 'demo';
 function DemoSeeder() {
   const { data: jobs } = useJobs();
   const createJob = useCreateJob();
-  const [seeded, setSeeded] = useState(false);
+  const seeded = useRef(false);
 
   useEffect(() => {
-    if (!isDemo || seeded || !jobs || jobs.length > 0) return;
+    if (!isDemo || seeded.current || !jobs || jobs.length > 0) return;
     demoJobs.forEach((job) => createJob.mutate(job));
-    setSeeded(true);
-  }, [jobs, seeded, createJob]);
+    seeded.current = true;
+  }, [jobs, createJob]);
 
   return null;
 }
@@ -24,8 +24,10 @@ function DemoSeeder() {
 export default function JobsPage() {
   return (
     <div className="flex flex-col h-full gap-5">
-      <div className="flex-shrink-0">
-        <h1 className="text-2xl font-bold text-gray-900 dark:text-gray-50">Job Tracker</h1>
+      <div className="shrink-0">
+        <h1 className="text-2xl font-bold text-gray-900 dark:text-gray-50">
+          Job Tracker
+        </h1>
         <p className="text-sm text-gray-500 dark:text-gray-400 mt-0.5">
           Track and manage all your job applications in one place.
         </p>
