+++ frontend/app/components/jobs/KanbanColumn.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useDroppable } from '@dnd-kit/core';
-import type { Job } from '../../lib/api';
+import type { Job } from '../../types/job';
 import JobCard from './JobCard';
 
 export interface ColumnConfig {
