+++ frontend/app/lib/api.ts
@@ -1,4 +1,7 @@
 import axios from 'axios';
+import type { Job, CreateJobInput, UpdateJobInput } from '../types/job';
+
+export type { Job, CreateJobInput, UpdateJobInput };
 
 const BASE_URL = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3001';
 
@@ -6,35 +9,6 @@ const apiClient = axios.create({
   baseURL: BASE_URL,
   headers: { 'Content-Type': 'application/json' },
 });
-
-export interface Job {
-  id: string;
-  company: string;
-  role: string;
-  status: string;
-  location: string;
-  salary: string;
-  jobUrl: string;
-  notes: string;
-  appliedAt: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface CreateJobInput {
-  company: string;
-  role: string;
-  status: string;
-  location?: string;
-  salary?: string;
-  jobUrl?: string;
-  notes?: string;
-}
-
-export interface UpdateJobInput {
-  id: string;
-  data: Partial<CreateJobInput>;
-}
 
 export const jobsApi = {
   getAll: (): Promise<Job[]> =>
