+++ frontend/app/types/job.ts
@@ -1,0 +1,28 @@
+export interface Job {
+  id: string;
+  company: string;
+  role: string;
+  status: string;
+  location: string;
+  salary: string;
+  jobUrl: string;
+  notes: string;
+  appliedAt: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface CreateJobInput {
+  company: string;
+  role: string;
+  status: string;
+  location?: string;
+  salary?: string;
+  jobUrl?: string;
+  notes?: string;
+}
+
+export interface UpdateJobInput {
+  id: string;
+  data: Partial<CreateJobInput>;
+}
